Cleanup and optimize the existing wallet flows (#2575)
* Cleanup and optimize the existing wallet flows

This doesn't add any new flows but dramatically optimizes the connect flow.
I realized that creating a web3 object is pretty expensive and we were doing this everytime
we instantiate a WalletAccessor. In this PR, we create one within the model lazily and use that
whenever we can. This makes sure that no UI opening up is blocked and shows the connect sheet and
the info about the dApp much much quicker. Also optimizes the tab opens from the Wallet Detail UI.

For sites that are not trusted, the flow requires the user to vouch for trust to show the connect button
otherwise we dismiss the sheet and dont connect.

* Add a temporary allowlist of marketplaces for easier daily driving

* Cleanup log

* Cleanup trust signal logic

Move all checks to a single place, so that we can swap that with a dedicated TrustHandler