Update node URLs and allow the user to pick the chain to connect through (#2631)
This is preliminary to support Polygon transactions. It starts plumbing the chain ID around
through transactions related communications so that it can be used later. It also adds a menu
to the connection UI so that the user can declare the network to connect through. This is currently
a must have since some dApps expect the user to declare willingness to connect over Polygon rather
than just communicate the right chainID through the API as a param.