Fix issue with Card transition animation not completing (#2769)
Switch from a spring animation to an easing animation as an experiment
to see if this eliminates the bug. The theory is that our logic for detecting
the end of the animation may not work properly with spring animations.
Perhaps the spring animation does not always precisely terminate at the
target value. Maybe there is a small rounding error sometimes. I'm not
100% convinced this is the answer, but it seems worth trying out.

Fixes #2181