Switch back to updated TabCardModel synchronously (#2860)
Since the CardDetails system uses IDs to refer to Tabs, its a problem
if TabCardModel and TabManager are ever out of sync in terms of what
tabs they know about.

So instead of updating TabCardModel asynchronously, we need to update
it synchronously. To achieve that, it's important for TabManager.tabs
to not be a published variable. Published variables notify observers
before they are modified, not after. TabCardModel wants to read from
the TabManager upon notification, so those notifications need to happen
after TabManager.tabs is updated.

This is a good thing as well because it sets the stage to allow us to
optimize a bit more. For bulk operations on TabManager, we can defer
notifications until all of the operations complete. This PR doesn't
really take advantage of that. Just adds some TODO comments about it.

With this change we can undo the guard added to protect buildRows from
accessing a TabCardDetails referencing a Tab that the TabManager no
longer knows about.