Ensure allDetails is always consistent with TabManager.tabs (#2991)
We do this by making allDetails not be @Published. That way
we can control when the TabCardModel signals any listeners. An
explicit objectWillChange.send() at the end of onDataUpdated
achieves this.

Also updates TabGroupCardModel in a similar fashion and cleans
up some code.

Reverts bandaid fix to handle a nil Tab from a TabCardDetails,
so we can see if this fix didn't work.

Fixes: #2964