Ensure switcher shows the correct state (#3203)
I've noticed that it is possible to not get an .onEnded notification
from a drag operation. I can repro by dragging around the top of the
device and causing the OS UI to kick in.

This change ensures that whenever we enter the tab switcher, the
drag offset of the switcher is reset. This way it will be updated to
show the correct state.

Also addresses an issue with `selectTab(nil)` that could result in
forcing the user out of incognito mode, which is not intended in our
UX design. We don't call `selectTab(nil)` so this was avoided, but
this change just helps remove this landmine for the future.

Fixes #3191