Plumbing to add apple oauth authentication code to /login-mobile calls (#3215)
After completing the native Apple oauth flow, we were previously only sending back the `identityToken` to the `/login-mobile` endpoint. We passed this as the `serverAuthCode` URL param.

With this change, we:

1. Rename the `serverAuthCode` URL param to `identityToken`, for clarity (the server API will still accept either param, preferring `identityToken` over `serverAuthCode` if both are set)
2. Add the authentication code to this request, as the new `authenticationCode` URL param.

The authentication code is used by httpd to exchange with Apple's servers for an authentication token, which is then encrypted and stored in mysql. A daily job (usertable-cleanauth) runs to verify the token has not been revoked by the user. See this doc for more details: https://docs.google.com/document/d/1F8Kwpexb481eIOBXxVuK9noBVCR8cHfZQR7HL6g06IA.

I also removed some legacy logic that was storing the identity token in a cookie, after verifying no code in httpd reads this cookie.