Add support for matching the initialURL of (non-)restored tabs (#3233)
addTabToTabGroupIfNeeded was only capable of matching against a
tab that was created during the current session. Once the app
was killed, on next cold start any non-restored tabs or even
tabs that had been restored would not be candidates for matching.

The underlying issue is our use of internal:// URLs to handle
session restore. When we check the WebView for its history item's
initialURL property, that could be an internal:// URL. Or if we
haven't restored the tab yet, it could be nil. We then need to
check sessionData.urls[0].

Fixes #3224